Working RBF GLobal, Global chunk, Dynamic Nearest Neigh
Working RBF GLobal, Global chunk, Dynamic Nearest Neigh
The nearest neighbour precomputed is not working nor the logistic or theory behind yet. The global and/or chunk is not computationally feasible for large data sets.
The dynamic nearest neighbours need to be computed on the fly, but it works well.